feat: abstracted oauth client
