+++ Shared/Tracking/Trackers/myanimelist/MyAnimeListTracker.swift
@@ -27,7 +27,7 @@ class MyAnimeListTracker: OAuthTracker {
     let api = MyAnimeListApi()
 
     let callbackHost = "myanimelist-auth"
-    lazy var authenticationUrl = api.authenticationUrl ?? ""
+    lazy var authenticationUrl = api.oauth.authenticationUrl ?? ""
 
     func register(trackId: String) async {
         guard let id = Int(trackId) else { return }
@@ -85,7 +85,7 @@ class MyAnimeListTracker: OAuthTracker {
 
     func handleAuthenticationCallback(url: URL) async {
         if let authCode = url.queryParameters?["code"] {
-            guard let oauth = await api.getAccessToken(authCode: authCode) else { return }
+            guard let oauth = await api.oauth.getAccessToken(authCode: authCode) else { return }
             token = oauth.accessToken
             UserDefaults.standard.set(try? JSONEncoder().encode(oauth), forKey: "Token.\(id).oauth")
         }
